refactor: extract uploadImage to a component
